refactor: move waveform classes into their family modules
Each Waveform subclass now lives next to its generator in
ripplegw/waveforms/<Family>.py and self-registers via @register; interfaces.py
holds only the Waveform ABC and the waveform_preset alias. __init__.py
re-exports the classes, so 'from ripplegw import IMRPhenomD' and waveform_preset
keep working. No behaviour change.
