+++ src/ripplegw/interfaces.py
@@ -1,21 +1,16 @@
+"""Abstract base class for ripple waveform models.
+
+The concrete waveform families live in their own modules under
+``ripplegw.waveforms`` (and other subpackages) and register themselves via
+``@register`` from :mod:`ripplegw.registry`. Construct any model through
+:func:`ripplegw.waveform`.
+"""
+
 from abc import ABC, abstractmethod
 
-import jax.numpy as jnp
 from jaxtyping import Array, Complex, Float
 
-from ripplegw.waveforms.TaylorF2 import gen_TaylorF2_hphc
-from ripplegw.waveforms.IMRPhenomD import gen_IMRPhenomD_hphc
-from ripplegw.waveforms.IMRPhenomD_NRTidalv2 import gen_IMRPhenomD_NRTidalv2_hphc
-from ripplegw.waveforms.IMRPhenomHM import gen_IMRPhenomHM
-from ripplegw.waveforms.IMRPhenomPv2 import gen_IMRPhenomPv2_hphc
-from ripplegw.waveforms.IMRPhenomXAS import gen_IMRPhenomXAS_hphc
-from ripplegw.waveforms.IMRPhenomXAS_NRTidalv3 import gen_IMRPhenomXAS_NRTidalv3_hphc
-from ripplegw.waveforms.IMRPhenomXHM import gen_IMRPhenomXHM_hphc
-from ripplegw.waveforms.IMRPhenomXP import gen_IMRPhenomXP_hphc
-from ripplegw.waveforms.IMRPhenomXPHM import generate_xphm
-from ripplegw.waveforms.SineGaussian import gen_SineGaussian_hphc
-from ripplegw.conversions import Mc_eta_to_ms
-from ripplegw.registry import WAVEFORM_REGISTRY, register
+from ripplegw.registry import WAVEFORM_REGISTRY
 
 
 class Waveform(ABC):
@@ -62,796 +57,9 @@ class Waveform(ABC):
         raise NotImplementedError("Waveform.__call__ must be implemented by subclasses")
 
 
-@register("TaylorF2", domain="FD", is_tidal=True, is_precessing=False)
-class TaylorF2(Waveform):
-    """TaylorF2 post-Newtonian frequency-domain waveform including tidal effects.
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-        use_lambda_tildes (bool): If True, expects ``lambda_tilde`` and
-            ``delta_lambda_tilde``; otherwise expects ``lambda_1`` and ``lambda_2``.
-    """
-
-    f_ref: float
-    use_lambda_tildes: bool
-
-    def __init__(self, f_ref: float = 20.0, use_lambda_tildes: bool = False) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-            use_lambda_tildes (bool): Whether to parameterise tidal deformability
-                via ``lambda_tilde`` / ``delta_lambda_tilde`` (as in Eq. 5-6 of
-                arXiv:1402.5156) instead of ``lambda_1`` / ``lambda_2``.
-                Defaults to False.
-        """
-        self.f_ref = f_ref
-        self.use_lambda_tildes = use_lambda_tildes
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            *(
-                ("lambda_tilde", "delta_lambda_tilde")
-                if self.use_lambda_tildes
-                else ("lambda_1", "lambda_2")
-            ),
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the TaylorF2 waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys ``M_c``,
-                ``eta``, ``s1_z``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``,
-                plus tidal keys depending on ``use_lambda_tildes``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        if self.use_lambda_tildes:
-            first_lambda_param = params["lambda_tilde"]
-            second_lambda_param = params["delta_lambda_tilde"]
-        else:
-            first_lambda_param = params["lambda_1"]
-            second_lambda_param = params["lambda_2"]
-
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_z"],
-                params["s2_z"],
-                first_lambda_param,
-                second_lambda_param,
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_TaylorF2_hphc(
-            frequency, theta, self.f_ref, use_lambda_tildes=self.use_lambda_tildes
-        )
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"TaylorF2(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomD", domain="FD", is_tidal=False, is_precessing=False)
-class IMRPhenomD(Waveform):
-    """IMRPhenomD frequency-domain waveform (non-precessing, aligned spins).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomD waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_z``, ``s2_z``, ``d_L``,
-                ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_z"],
-                params["s2_z"],
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomD_hphc(frequency, theta, self.f_ref)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomD(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomD_NRTidalv2", domain="FD", is_tidal=True, is_precessing=False)
-class IMRPhenomD_NRTidalv2(Waveform):
-    """IMRPhenomD_NRTidalv2 frequency-domain waveform (non-precessing, NRTidalv2 tides).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-        use_lambda_tildes (bool): If True, expects ``lambda_tilde`` /
-            ``delta_lambda_tilde``; otherwise ``lambda_1`` / ``lambda_2``.
-        no_taper (bool): If True, the Planck taper in the amplitude is disabled.
-    """
-
-    f_ref: float
-    use_lambda_tildes: bool
-    no_taper: bool
-
-    def __init__(
-        self,
-        f_ref: float = 20.0,
-        use_lambda_tildes: bool = False,
-        no_taper: bool = False,
-    ) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-            use_lambda_tildes (bool): Whether to parameterise tidal deformability
-                via ``lambda_tilde`` / ``delta_lambda_tilde`` (Eq. 5-6 of
-                arXiv:1402.5156) instead of ``lambda_1`` / ``lambda_2``.
-                Defaults to False.
-            no_taper (bool): Whether to remove the Planck taper in the amplitude
-                (useful for relative binning runs). Defaults to False.
-        """
-        self.f_ref = f_ref
-        self.use_lambda_tildes = use_lambda_tildes
-        self.no_taper = no_taper
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            *(
-                ("lambda_tilde", "delta_lambda_tilde")
-                if self.use_lambda_tildes
-                else ("lambda_1", "lambda_2")
-            ),
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomD_NRTidalv2 waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys ``M_c``,
-                ``eta``, ``s1_z``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``,
-                plus tidal keys depending on ``use_lambda_tildes``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        if self.use_lambda_tildes:
-            first_lambda_param = params["lambda_tilde"]
-            second_lambda_param = params["delta_lambda_tilde"]
-        else:
-            first_lambda_param = params["lambda_1"]
-            second_lambda_param = params["lambda_2"]
-
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_z"],
-                params["s2_z"],
-                first_lambda_param,
-                second_lambda_param,
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomD_NRTidalv2_hphc(
-            frequency,
-            theta,
-            self.f_ref,
-            use_lambda_tildes=self.use_lambda_tildes,
-            no_taper=self.no_taper,
-        )
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomD_NRTidalv2(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomHM", domain="FD", is_tidal=False, is_precessing=False)
-class IMRPhenomHM(Waveform):
-    """IMRPhenomHM frequency-domain waveform (aligned spins, higher-order modes).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        output = {}
-        m1, m2 = Mc_eta_to_ms(jnp.array([params["M_c"], params["eta"]]))
-        hp, hc = gen_IMRPhenomHM(
-            frequency,
-            m1,
-            m2,
-            params["s1_z"],
-            params["s2_z"],
-            params["d_L"],
-            params["iota"],
-            params["phase_c"],
-            self.f_ref,
-        )
-        output["p"] = hp
-        output["c"] = hc
-        return output
-
-    def __repr__(self):
-        return f"IMRPhenomHM(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomPv2", domain="FD", is_tidal=False, is_precessing=True)
-class IMRPhenomPv2(Waveform):
-    """IMRPhenomPv2 frequency-domain waveform (precessing spins).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_x",
-            "s1_y",
-            "s1_z",
-            "s2_x",
-            "s2_y",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomPv2 waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_x``, ``s1_y``, ``s1_z``,
-                ``s2_x``, ``s2_y``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_x"],
-                params["s1_y"],
-                params["s1_z"],
-                params["s2_x"],
-                params["s2_y"],
-                params["s2_z"],
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomPv2_hphc(frequency, theta, self.f_ref)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomPv2(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomXAS", domain="FD", is_tidal=False, is_precessing=False)
-class IMRPhenomXAS(Waveform):
-    """IMRPhenomXAS frequency-domain waveform (non-precessing, aligned spins, X family).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomXAS waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_z``, ``s2_z``, ``d_L``,
-                ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_z"],
-                params["s2_z"],
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomXAS_hphc(frequency, theta, self.f_ref)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomXAS(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomXAS_NRTidalv3", domain="FD", is_tidal=True, is_precessing=False)
-class IMRPhenomXAS_NRTidalv3(Waveform):
-    """IMRPhenomXAS_NRTidalv3 frequency-domain waveform (non-precessing, NRTidalv3 tides).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-        use_lambda_tildes (bool): If True, expects ``lambda_tilde`` /
-            ``delta_lambda_tilde``; otherwise ``lambda_1`` / ``lambda_2``.
-        no_taper (bool): If True, the Planck taper in the amplitude is disabled.
-    """
-
-    f_ref: float
-    use_lambda_tildes: bool
-    no_taper: bool
-
-    def __init__(
-        self,
-        f_ref: float = 20.0,
-        use_lambda_tildes: bool = False,
-        no_taper: bool = False,
-    ) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-            use_lambda_tildes (bool): Whether to parameterise tidal deformability
-                via ``lambda_tilde`` / ``delta_lambda_tilde`` rather than
-                ``lambda_1`` / ``lambda_2``. Defaults to False.
-            no_taper (bool): Whether to disable tapering (useful for relative
-                binning runs). Defaults to False.
-        """
-        self.f_ref = f_ref
-        self.use_lambda_tildes = use_lambda_tildes
-        self.no_taper = no_taper
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            *(
-                ("lambda_tilde", "delta_lambda_tilde")
-                if self.use_lambda_tildes
-                else ("lambda_1", "lambda_2")
-            ),
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomXAS_NRTidalv3 waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys ``M_c``,
-                ``eta``, ``s1_z``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``,
-                plus tidal keys depending on ``use_lambda_tildes``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        if self.use_lambda_tildes:
-            first_lambda_param = params["lambda_tilde"]
-            second_lambda_param = params["delta_lambda_tilde"]
-        else:
-            first_lambda_param = params["lambda_1"]
-            second_lambda_param = params["lambda_2"]
-
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_z"],
-                params["s2_z"],
-                first_lambda_param,
-                second_lambda_param,
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomXAS_NRTidalv3_hphc(
-            frequency,
-            theta,
-            self.f_ref,
-            use_lambda_tildes=self.use_lambda_tildes,
-            no_taper=self.no_taper,
-        )
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomXAS_NRTidalv3(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomXHM", domain="FD", is_tidal=False, is_precessing=False)
-class IMRPhenomXHM(Waveform):
-    """IMRPhenomXHM frequency-domain waveform (aligned spins, higher-order modes).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_z",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomXHM waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_z``, ``s2_z``, ``d_L``,
-                ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        m1, m2 = Mc_eta_to_ms(jnp.array([params["M_c"], params["eta"]]))
-        theta = jnp.array(
-            [
-                m1,
-                m2,
-                params["s1_z"],
-                params["s2_z"],
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomXHM_hphc(frequency, theta, self.f_ref)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomXHM(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomXP", domain="FD", is_tidal=False, is_precessing=True)
-class IMRPhenomXP(Waveform):
-    """IMRPhenomXP frequency-domain waveform (precessing spins, 22-mode only).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_x",
-            "s1_y",
-            "s1_z",
-            "s2_x",
-            "s2_y",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomXP waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_x``, ``s1_y``, ``s1_z``,
-                ``s2_x``, ``s2_y``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        theta = jnp.array(
-            [
-                params["M_c"],
-                params["eta"],
-                params["s1_x"],
-                params["s1_y"],
-                params["s1_z"],
-                params["s2_x"],
-                params["s2_y"],
-                params["s2_z"],
-                params["d_L"],
-                0.0,
-                params["phase_c"],
-                params["iota"],
-            ]
-        )
-        hp, hc = gen_IMRPhenomXP_hphc(frequency, theta, self.f_ref)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomXP(f_ref={self.f_ref})"
-
-
-@register("IMRPhenomXPHM", domain="FD", is_tidal=False, is_precessing=True)
-class IMRPhenomXPHM(Waveform):
-    """IMRPhenomXPHM frequency-domain waveform (precessing spins, higher-order modes).
-
-    Attributes:
-        f_ref (float): Reference frequency in Hz.
-    """
-
-    f_ref: float
-
-    def __init__(self, f_ref: float = 20.0) -> None:
-        """
-        Args:
-            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
-        """
-        self.f_ref = f_ref
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return (
-            "M_c",
-            "eta",
-            "s1_x",
-            "s1_y",
-            "s1_z",
-            "s2_x",
-            "s2_y",
-            "s2_z",
-            "d_L",
-            "phase_c",
-            "iota",
-        )
-
-    def __call__(
-        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
-    ) -> dict[str, Complex[Array, " n_freq"]]:
-        """Evaluate the IMRPhenomXPHM waveform.
-
-        Args:
-            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
-            params (dict[str, Float]): Source parameters with keys
-                ``M_c``, ``eta``, ``s1_x``, ``s1_y``, ``s1_z``,
-                ``s2_x``, ``s2_y``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``.
-
-        Returns:
-            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        m1, m2 = Mc_eta_to_ms(jnp.array([params["M_c"], params["eta"]]))
-        hp, hc = generate_xphm(
-            m1,
-            m2,
-            params["s1_x"],
-            params["s1_y"],
-            params["s1_z"],
-            params["s2_x"],
-            params["s2_y"],
-            params["s2_z"],
-            params["d_L"],
-            params["iota"],
-            params["phase_c"],
-            frequency,
-            self.f_ref,
-        )
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return f"IMRPhenomXPHM(f_ref={self.f_ref})"
-
-
-@register("SineGaussian", domain="TD", is_tidal=False, is_precessing=False)
-class SineGaussian(Waveform):
-    """Sine-Gaussian time-domain burst waveform."""
-
-    def __init__(self) -> None:
-        pass
-
-    @property
-    def parameter_names(self) -> tuple[str, ...]:
-        return ("Q", "f_0", "hrss", "phase", "e")
-
-    def __call__(
-        self, t: Float[Array, " n_time"], params: dict[str, Float]
-    ) -> dict[str, Float[Array, " n_time"]]:
-        """Evaluate the SineGaussian waveform.
-
-        Args:
-            t (Float[Array, " n_time"]): Time grid centered at t=0. Create using
-                ``jnp.arange(-duration/2, duration/2, 1/fs)``.
-            params (dict[str, Float]): Source parameters with keys ``Q``
-                (quality factor), ``f_0`` (central frequency in Hz), ``hrss``,
-                ``phase``, ``e`` (eccentricity).
-
-        Returns:
-            dict[str, Float[Array, " n_time"]]: Plus (``"p"``) and cross (``"c"``)
-                polarizations.
-        """
-        theta = jnp.array(
-            [
-                params["Q"],
-                params["f_0"],
-                params["hrss"],
-                params["phase"],
-                params["e"],
-            ]
-        )
-        hp, hc = gen_SineGaussian_hphc(t, theta)
-        return {"p": hp, "c": hc}
-
-    def __repr__(self):
-        return "SineGaussian()"
-
-
 #: Mapping from model name strings to ``Waveform`` subclasses.
 #:
-#: This is an alias of the global :data:`ripplegw.registry.WAVEFORM_REGISTRY`,
-#: which every model populates via the :func:`~ripplegw.registry.register`
-#: decorator at class-definition time. Retained under this name for backward
-#: compatibility; prefer :func:`ripplegw.waveform` / :func:`ripplegw.list_waveforms`.
+#: Alias of :data:`ripplegw.registry.WAVEFORM_REGISTRY`, populated by every
+#: model via ``@register`` at import. Retained for backward compatibility;
+#: prefer :func:`ripplegw.waveform` / :func:`ripplegw.list_waveforms`.
 waveform_preset: dict[str, type[Waveform]] = WAVEFORM_REGISTRY
