+++ src/ripplegw/waveforms/IMRPhenomXP.py
@@ -1,8 +1,10 @@
 import jax
+from ripplegw.interfaces import Waveform
+from ripplegw.registry import register
+from ripplegw.conversions import Mc_eta_to_ms
 import jax.numpy as jnp
 from jaxtyping import Array, Float, Complex
 from ripplegw.constants import PI, MTSUN, MRSUN, MPC
-from ripplegw.conversions import Mc_eta_to_ms
 from ripplegw.waveforms import LALSimIMRPhenomX_precession as pPrec
 from ripplegw.waveforms.initialize_MSA_system import IMRPhenomX_Initialize_MSA_System
 
@@ -185,3 +187,74 @@ def gen_IMRPhenomXP_hphc(
     hp, hc = apply_polarization_rotation(zeta_polarisations, _hp, _hc)
 
     return hp, hc
+
+
+@register("IMRPhenomXP", domain="FD", is_tidal=False, is_precessing=True)
+class IMRPhenomXP(Waveform):
+    """IMRPhenomXP frequency-domain waveform (precessing spins, 22-mode only).
+
+    Attributes:
+        f_ref (float): Reference frequency in Hz.
+    """
+
+    f_ref: float
+
+    def __init__(self, f_ref: float = 20.0) -> None:
+        """
+        Args:
+            f_ref (float): Reference frequency in Hz. Defaults to 20.0.
+        """
+        self.f_ref = f_ref
+
+    @property
+    def parameter_names(self) -> tuple[str, ...]:
+        return (
+            "M_c",
+            "eta",
+            "s1_x",
+            "s1_y",
+            "s1_z",
+            "s2_x",
+            "s2_y",
+            "s2_z",
+            "d_L",
+            "phase_c",
+            "iota",
+        )
+
+    def __call__(
+        self, frequency: Float[Array, " n_freq"], params: dict[str, Float]
+    ) -> dict[str, Complex[Array, " n_freq"]]:
+        """Evaluate the IMRPhenomXP waveform.
+
+        Args:
+            frequency (Float[Array, " n_freq"]): Frequency array in Hz.
+            params (dict[str, Float]): Source parameters with keys
+                ``M_c``, ``eta``, ``s1_x``, ``s1_y``, ``s1_z``,
+                ``s2_x``, ``s2_y``, ``s2_z``, ``d_L``, ``phase_c``, ``iota``.
+
+        Returns:
+            dict[str, Complex[Array, " n_freq"]]: Plus (``"p"``) and cross (``"c"``)
+                polarizations.
+        """
+        theta = jnp.array(
+            [
+                params["M_c"],
+                params["eta"],
+                params["s1_x"],
+                params["s1_y"],
+                params["s1_z"],
+                params["s2_x"],
+                params["s2_y"],
+                params["s2_z"],
+                params["d_L"],
+                0.0,
+                params["phase_c"],
+                params["iota"],
+            ]
+        )
+        hp, hc = gen_IMRPhenomXP_hphc(frequency, theta, self.f_ref)
+        return {"p": hp, "c": hc}
+
+    def __repr__(self):
+        return f"IMRPhenomXP(f_ref={self.f_ref})"
